cloud-functions/webcrawler_isc/requirements.txt
psycopg2>=2.9.10,<2.10
python-dotenv